requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
fastapi==0.110.0
uvicorn==0.29.0
psycopg==3.1.19
sqlalchemy==2.0.39
alembic==1.13.1
pydantic==2.10.6
greenlet==3.0.3